LPS-111561 Isolate Vertical Alignment selector
